eris files refactor, #160
