+++ services/manage.go
@@ -11,6 +11,7 @@ import (
 
 	"github.com/eris-ltd/eris-cli/data"
 	"github.com/eris-ltd/eris-cli/definitions"
+	"github.com/eris-ltd/eris-cli/ipfs"
 	"github.com/eris-ltd/eris-cli/loaders"
 	"github.com/eris-ltd/eris-cli/perform"
 	"github.com/eris-ltd/eris-cli/util"
@@ -31,9 +32,9 @@ func ImportService(do *definitions.Do) error {
 
 		var err error
 		if logger.Level > 0 {
-			err = util.GetFromIPFS(s[1], fileName, "", logger.Writer)
+			err = ipfs.GetFromIPFS(s[1], fileName, "", logger.Writer)
 		} else {
-			err = util.GetFromIPFS(s[1], fileName, "", bytes.NewBuffer([]byte{}))
+			err = ipfs.GetFromIPFS(s[1], fileName, "", bytes.NewBuffer([]byte{}))
 		}
 
 		if err != nil {
@@ -303,9 +304,9 @@ func exportFile(servName string) (string, error) {
 	var hash string
 	var err error
 	if logger.Level > 0 {
-		hash, err = util.SendToIPFS(fileName, false, logger.Writer)
+		hash, err = ipfs.SendToIPFS(fileName, false, logger.Writer)
 	} else {
-		hash, err = util.SendToIPFS(fileName, false, bytes.NewBuffer([]byte{}))
+		hash, err = ipfs.SendToIPFS(fileName, false, bytes.NewBuffer([]byte{}))
 	}
 
 	if err != nil {
